RHOAIENG-79891: Extract ConnectionDetailsHelperText, DefaultValueTextRenderer, FieldMode to ui-core
