+++ frontend/src/concepts/connectionTypes/fields/DropdownFormField.tsx
@@ -17,9 +17,9 @@ import {
   SelectOption,
 } from '@patternfly/react-core';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
+import DefaultValueTextRenderer from '@odh-dashboard/ui-core/components/connectionTypes/DefaultValueTextRenderer';
 import { DropdownField } from '#~/concepts/connectionTypes/types';
 import { FieldProps } from '#~/concepts/connectionTypes/fields/types';
-import DefaultValueTextRenderer from '#~/concepts/connectionTypes/fields/DefaultValueTextRenderer';
 
 const DropdownFormField: React.FC<FieldProps<DropdownField>> = ({
   id,
