+++ frontend/src/concepts/connectionTypes/fields/TextFormField.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { TextArea } from '@patternfly/react-core';
 import { trimInputOnBlur, trimInputOnPaste } from '@odh-dashboard/ui-core/utilities';
+import DefaultValueTextRenderer from '@odh-dashboard/ui-core/components/connectionTypes/DefaultValueTextRenderer';
 import { TextField } from '#~/concepts/connectionTypes/types';
 import { FieldProps } from '#~/concepts/connectionTypes/fields/types';
-import DefaultValueTextRenderer from '#~/concepts/connectionTypes/fields/DefaultValueTextRenderer';
 
 const TextFormField: React.FC<FieldProps<TextField>> = ({
   id,
