+++ packages/model-serving/src/components/deploymentWizard/fields/modelLocationFields/ExistingConnectionField.tsx
@@ -12,7 +12,7 @@ import type { Connection, ConnectionTypeConfigMapObj } from '@odh-dashboard/k8s-
 import TypeaheadSelect, {
   TypeaheadSelectOption,
 } from '@odh-dashboard/ui-core/components/TypeaheadSelect';
-import { ConnectionDetailsHelperText } from '@odh-dashboard/internal/concepts/connectionTypes/ConnectionDetailsHelperText';
+import { ConnectionDetailsHelperText } from '@odh-dashboard/ui-core/components/connectionTypes/ConnectionDetailsHelperText';
 import ConnectionS3FolderPathField from '@odh-dashboard/internal/pages/modelServing/screens/projects/InferenceServiceModal/ConnectionS3FolderPathField';
 import ConnectionOciPathField from '@odh-dashboard/internal/pages/modelServing/screens/projects/InferenceServiceModal/ConnectionOciPathField';
 import {
